Major Refactoring, moved InteractiveSever to core
- Code cleanup, moved thymeleaf to core
  completely, junit calls to core for html report
- Move examples Dataclasses to core.
- Remove unneeded Ktor deps from application.
