+++ application/build.gradle
@@ -78,16 +78,10 @@ dependencies {
     implementation(project(':junit5-support'))
 
     implementation "io.ktor:ktor-client-cio:$ktor_version"
+    implementation "io.ktor:ktor-server-core:$ktor_version"
     implementation 'io.swagger.parser.v3:swagger-parser:2.1.22'
 
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json-jvm:1.6.3"
-
-    implementation "io.ktor:ktor-server-netty:$ktor_version"
-    implementation "io.ktor:ktor-server-core:$ktor_version"
-    implementation "io.ktor:ktor-server-cors:$ktor_version"
-    implementation("io.ktor:ktor-server-content-negotiation:$ktor_version")
-    implementation("io.ktor:ktor-server-status-pages:$ktor_version")
-    implementation "io.ktor:ktor-serialization-jackson:$ktor_version"
 
     testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:$junit_version"
