+++ core/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     implementation("io.ktor:ktor-server-double-receive:$ktor_version")
     implementation("io.ktor:ktor-server-content-negotiation:$ktor_version")
     implementation "io.ktor:ktor-serialization-jackson:$ktor_version"
+    implementation("io.ktor:ktor-server-status-pages:$ktor_version")
     implementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.18.0'
