+++ core/src/main/kotlin/io/specmatic/templates/HtmlTemplateConfiguration.kt
@@ -1,4 +1,4 @@
-package io.specmatic.test.reports.coverage.html
+package io.specmatic.templates
 import org.thymeleaf.TemplateEngine
 import org.thymeleaf.context.Context
 import org.thymeleaf.templatemode.TemplateMode
@@ -6,7 +6,7 @@ import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver
 
 class HtmlTemplateConfiguration {
     companion object {
-        fun configureTemplateEngine(): TemplateEngine {
+        private fun configureTemplateEngine(): TemplateEngine {
             val templateResolver = ClassLoaderTemplateResolver().apply {
                 prefix = "templates/"
                 suffix = ".html"
