+++ core/src/main/resources/templates/example/index.html
