+++ junit5-support/build.gradle
@@ -25,7 +25,6 @@ dependencies {
     implementation 'org.assertj:assertj-core:3.26.3'
     implementation 'org.junit.jupiter:junit-jupiter-api:5.11.2'
     implementation 'com.fasterxml.jackson.core:jackson-databind:2.18.0'
-    implementation 'org.thymeleaf:thymeleaf:3.1.2.RELEASE'
 
     implementation "io.ktor:ktor-client-core-jvm:${ktor_version}"
     implementation "io.ktor:ktor-client-cio:${ktor_version}"
